+++ CHANGELOG.md
@@ -1,14 +1,24 @@
 # Changelog
 
-## 1.0.0
+## 1.1.3
 
-Initial release.
+- Support referencing other variables in variable queries
+- Fix Query input field overflow issue when content is too long
+- Optimize "All" option handling: expand to actual values instead of regex
+
+## 1.1.2
+
+Fix interval issue.
+
+## 1.1.1
+
+Fix Bug.
 
 ## 1.1.0
 
 Add promql support.
 Add Legend support.
 
-## 1.1.1
+## 1.0.0
 
-Fix interval issue.
+Initial release.